feat: Improve Escape Tools
